modal changed to ActivateAccountPage.jsx, app router changed.
