+++ financiation-react/src/App.js
@@ -13,7 +13,6 @@ import AddVisitPage from "./pages/AddVisitPage";
 import VisitRegisterPage from "./pages/VisitRegisterPage";
 import PrivateRoute from "./utils/PrivateRoute";
 import {AuthProvider} from "./context/AuthContext";
-import Successful from "./components/Successful";
 import GroupCard from "./components/GroupCard";
 import FormPage from "./pages/FormPage";
 
@@ -31,7 +30,6 @@ function App() {
                         <Route path='/reset-password' element={<ResetPasswordPage/>}/>
                         <Route path='/reset-password/confirm' element={<ResetPasswordConfirmPage/>}/>
                         <Route path='/activate' element={<ActivateAccountPage/>}/>
-                        <Route exact path='/successful' element={<Successful/>}/>
 
                         <Route path='/advised' exact element={<PrivateRoute><AdvisedListPage/></PrivateRoute>}/>
                         <Route path='/advised/:id' element={<PrivateRoute><AdvisedPage/></PrivateRoute>}/>
