+++ financiation-react/src/pages/ActivateAccountPage.jsx
@@ -1,8 +1,13 @@
 import React from "react";
-import {Button, Card, Col, Container, Row} from "react-bootstrap";
+import {Button, Card, Col, Container, Modal, Row} from "react-bootstrap";
 import verifyimg from '../assets/images/verifyimg.gif';
+import Check from "../assets/images/checked.gif";
 
 const ActivateAccountPage = () => {
+
+    const [show, setShow] = React.useState(false);
+    const handleClose = () => setShow(false);
+    const handleShow = () => setShow(true);
     return (
         <div className='container mt-5'>
 
@@ -20,18 +25,34 @@ const ActivateAccountPage = () => {
                                 <div className='py-2'>
                                     <Row>
                                         <Col>
-                                            <Button>Verificar</Button>
+                                            <Button onClick={handleShow}>Verificar</Button>
                                         </Col>
                                     </Row>
                                 </div>
+
+                                <Modal show={show} onHide={handleClose}>
+                                    <Modal.Header closeButton>
+                                        <Modal.Title id="Titulo">Verificacion de cuenta</Modal.Title>
+                                    </Modal.Header>
+
+                                    <Modal.Body>
+                                        <img src={Check} alt="CheckButton" id="CheckButton" className="mx-auto img-fluid"/>
+                                        <p className="text-center">¡Su cuenta ha sido verficada!</p>
+
+                                    </Modal.Body>
+                                    <Modal.Footer>
+                                        <Button variant="success" onClick={handleClose}>
+                                            OK
+                                        </Button>
+                                    </Modal.Footer>
+                                </Modal>
                             </Card.Body>
                         </Card>
                     </Col>
                 </Row>
             </Container>
-
         </div>
-    )
+    );
 }
 
 export default ActivateAccountPage
